Close 3 bugs:
Close #11: Update all get_<geo>_demographics functions to say they are returning the data they are actually returning.

Close #13: Update all get_<geo>_demographics functions to remove references which are no longer valid.

Close #15: The example for get_county_demographics was failing. The issue is that tidycensus returns fips codes as characters with a leading "0", but choroplethr expects them as a number (with no leading "0")