Remove function choroplethr_animate. As part of the unarchiving process CRAN has manually reviewed the package and found the following issues with that function. I simply don't want to invest the time in addressing them now, so am opting to delete the function instead. It can always be brought back later if someone wants.
You write information messages to the console that cannot be easily
suppressed.
It is more R like to generate objects that can be used to extract the
information a user is interested in, and then print() that object.
Instead of print() rather use message()/warning() or
if(verbose)print(..) (or maybe stop()) if you really have to write text
to the console. (except for print, summary, interactive functions)
-> R/choroplethr_animate.R
For more details:
<https://contributor.r-project.org/cran-cookbook/code_issues.html#using-printcat>

Please ensure that your functions do not write by default or in your
examples/vignettes/tests in the user's home filespace (including the
package directory and getwd()). This is not allowed by CRAN policies.
Please omit any default path in writing functions. In your
examples/vignettes/tests you can write to tempdir().
-> R/choroplethr_animate.R
For more details:
<https://contributor.r-project.org/cran-cookbook/code_issues.html#writing-files-and-directories-to-the-home-filespace>

Please always make sure to reset to user's options(), working directory
or par() after you changed it in examples and vignettes and demos. ->
man/choroplethr_animate.Rd
e.g.:
old <- setwd(...)
...
setwd(old)
For more details:
<https://contributor.r-project.org/cran-cookbook/code_issues.html#change-of-options-graphical-parameters-and-working-directory>